move macro id to ids
