+++ crates/ra_hir/src/lib.rs
@@ -42,8 +42,8 @@ pub use self::{
     path::{Path, PathKind},
     name::Name,
     krate::Crate,
-    ids::{HirFileId, DefId, DefLoc},
-    macros::{MacroDef, MacroInput, MacroExpansion, MacroCallId, MacroCallLoc},
+    ids::{HirFileId, DefId, DefLoc, MacroCallId, MacroCallLoc},
+    macros::{MacroDef, MacroInput, MacroExpansion},
     module::{Module, ModuleId, Problem, nameres::{ItemMap, PerNs, Namespace}, ModuleScope, Resolution},
     function::{Function, FnScopes},
     adt::{Struct, Enum},
